+++ controllers/about/index.js
@@ -1,13 +1,14 @@
 'use strict';
 const express = require('express');
 const router = express.Router();
-const routeStatic = require('../utils/routeStatic');
+const routeStatic = require('../routeStatic');
 
 module.exports = (pages, sectionPath, sectionId) => {
-    /**
-     * 1. Populate static pages
-     */
-    routeStatic.initRouting(pages, router, sectionPath, sectionId);
-
+    routeStatic.init({
+        router: router,
+        pages: pages,
+        sectionPath: sectionPath,
+        sectionId: sectionId
+    });
     return router;
 };
